added dropdown on navBar, implemented logout button
