+++ src/client/Components/Home/Home.tsx
@@ -3,35 +3,23 @@ import { Link, useNavigate } from 'react-router-dom';
 import { useDispatch, useSelector } from 'react-redux';
 
 import { IReducers } from '../../Interfaces/IReducers';
-import { signIn } from '../../Store/actions';
 import './styles.css'
+import NavBar from './NavBar'
 
 const Home = () => {
   const appReducer = useSelector((state: IReducers) => state.appReducer);
-  const dispatch = useDispatch();
-  const navigate = useNavigate();
 
   useEffect(() => {
-    console.log('signInState: ', appReducer.signInState)
-    console.log('Signed in username: ', appReducer.username)
-    console.log('JWT token stored in localStorage: ', localStorage.getItem('token'));
+    console.log('signInState from store: ', appReducer.signInState)
+    console.log('Signed in username from localStorage: ', localStorage.getItem('username'))
+    console.log('JWT token stored from localStorage: ', localStorage.getItem('token'));
   }, [appReducer]);
-
-  const handleLogOut = (): void => {
-    localStorage.removeItem('token');
-    localStorage.removeItem('username');
-    dispatch(signIn({
-      signInState: false,
-      username: localStorage.getItem('username')
-    }));
-    navigate('/');
-  }
 
   return (
     <div>
-      <div><span>NavBar placeholder</span><button className="btn" id='logout-btn'onClick={handleLogOut}>Logout</button></div>
+      <NavBar />
       This is the home page.
-      <div>{appReducer.username}</div>
+      <div>{'username: ' + localStorage.getItem('username')}</div>
     </div>
   )
 }
