+++ src/client/Components/Home/styles.css
@@ -1,3 +1,14 @@
 #logout-btn {
   float: right
 }
+
+#navbar {
+  padding: 0.5rem;
+  margin-bottom: 1rem;
+  border: 1px solid green;
+}
+
+#dropdown {
+  float: right;
+  margin-right: 1rem;
+}
